+++ CHANGELOG.md
@@ -10,5 +10,7 @@
 
   Akin to `SmartIdSession`, you can use `MobileIdSession.serialize` and `MobileIdSession.parse` to save the session somewhere.
 
+- Renames `MobileId.prototype.readCertificate` to `MobileId.prototype.certificate` to match the `SmartId` API.
+
 ## 0.1.337 (Nov 10, 2023)
 - Sans three-headed dogs.